// packages/skeleton-angular/projects/lib/src/lib/components/navigation/types.ts
export enum NavigationParent {
  Bar = 'bar',
  Rail = 'rail',
  None = 'none',
}
export interface NavContext {
  parent: Signal<string>;
  value: Signal<string>;
  expanded: Signal<boolean>;
  onSelectionHandler: (id: string) => void;
}

// Components ---

/* Props common between NavRail and NavBar */
export interface NavCommonProps {
  value?: string;

  // Root ---
  /** Set base styles. */
  base?: string;
  /** Set background classes. */
  background?: string;
  /** Set padding classes. */
  padding?: string;
  /** Set width classes. */
  width?: string;
  /** Set width classes for expanded mode. */
  widthExpanded?: string;
  /** Set width classes. */
  height?: string;
  /** Provide arbitrary CSS classes. */
  classes?: string;

  // Tiles ---
  /** Set base classes. */
  tilesBase?: string;
  /** Set flex direction classes. */
  tilesFlexDirection?: string;
  /** Set flex justify classes. */
  tilesJustify?: string;
  /** Set flex align classes. */
  tilesItems?: string;
  /** Set gap classes. */
  tilesGap?: string;
  /** Provide arbitrary CSS classes. */
  tilesClasses?: string;

  // Events ---
  /** Triggers when selection occurs. */
  // onchange?: (id: string) => void;
}

export interface NavBarProps extends NavCommonProps {
  // Snippets ---
  /** The default children snippet. */
  // children?: Snippet;
}

export interface NavRailProps extends NavCommonProps {
  /** Enabled expanded mode. */
  expanded?: boolean;

  // Header ---
  /** Set base classes. */
  headerBase?: string;
  /** Set flex direction classes. */
  headerFlexDirection?: string;
  /** Set flex justify classes. */
  headerJustify?: string;
  /** Set flex align classes. */
  headerItems?: string;
  /** Set gap classes. */
  headerGap?: string;
  /** Provide arbitrary CSS classes. */
  headerClasses?: string;

  // Footer ---
  /** Set base classes. */
  footerBase?: string;
  /** Set flex direction classes. */
  footerFlexDirection?: string;
  /** Set flex justify classes. */
  footerJustify?: string;
  /** Set flex align classes. */
  footerItems?: string;
  /** Set gap classes. */
  footerGap?: string;
  /** Provide arbitrary CSS classes. */
  footerClasses?: string;

  // // Snippets ---
  // /** The header snippet. */
  // header?: Snippet;
  // /** The tiles snippet. */
  // tiles?: Snippet;
  // /** The footer snippet. */
  // footer?: Snippet;
}